+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Tue, 14 Feb 2023 15:44:48 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Feb 2023 17:27:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.1.11
+
+Tue, 14 Feb 2023 17:27:05 GMT
+
+### Patches
+
+- Add column IDs to table ([ni/nimble@a88731d](https://github.com/ni/nimble/commit/a88731d393687500fc23f86437a73bc6c947e17f))
 
 ## 11.1.9